requirements.txt
Flask
python-dotenv
google-generativeai
google-api-python-client
PyMuPDF
python-docx
numpy
scikit-learn
google-cloud-aiplatform
huggingface-hub